Caching duration should not require a [long], having an [int] with roughly 2^31 seconds (~ 68 years) should be enough. Especially since RFC-2616 suggests caching should not exceed one year.
This will adjust to IResourceSettings#getDefaultCacheDuration() which already was an [int].

git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@996118 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket/src/main/java/org/apache/wicket/protocol/http/RequestUtils.java
@@ -309,7 +309,7 @@ public final class RequestUtils
 	 *
 	 * @see RequestUtils#MAX_CACHE_DURATION
 	 */
-	public static void enableCaching(WebResponse response, long duration, boolean cachePublic)
+	public static void enableCaching(WebResponse response, int duration, boolean cachePublic)
 	{
 		Args.notNull(response, "response");
